requirements.txt
aio-pika==9.4.0
aiohttp==3.9.1
aiormq==6.8.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
asyncpg==0.29.0
asyncpg-stubs==0.29.1
attrs==23.2.0
backoff==2.2.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cloudpickle==3.0.0
colorama==0.4.6
contextlib2==21.6.0
croniter==2.0.1
cryptography==42.0.1
Deprecated==1.2.14
executing==2.0.1
frozenlist==1.4.1
googleapis-common-protos==1.62.0
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
icecream==2.1.3
idna==3.6
importlib-metadata==6.11.0
marshmallow==3.20.2
marshmallow-dataclass==8.6.0
multidict==6.0.4
mypy==1.8.0
mypy-extensions==1.0.0
names==0.3.0
opentelemetry-api==1.22.0
opentelemetry-distro==0.43b0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-exporter-prometheus==0.43b0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-aio-pika==0.43b0
opentelemetry-instrumentation-aiohttp-client==0.43b0
opentelemetry-instrumentation-aws-lambda==0.43b0
opentelemetry-instrumentation-dbapi==0.43b0
opentelemetry-instrumentation-logging==0.43b0
opentelemetry-instrumentation-sqlite3==0.43b0
opentelemetry-instrumentation-urllib==0.43b0
opentelemetry-instrumentation-wsgi==0.43b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
packaging==23.2
pamqp==3.3.0
prometheus-client==0.19.0
protobuf==4.25.2
psycopg2==2.9.9
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
returns==0.22.0
schema==0.7.5
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
sqlmodel==0.0.14
stateless==0.5.1
surrealdb==0.3.2
types-croniter==2.0.0.20240106
types-pyOpenSSL==23.3.0.20240106
types-redis==4.6.0.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
websockets==10.4
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
